requirements.txt
# Core Dependencies - Python 3.11 compatible
yt-dlp>=2023.10.13
faster-whisper>=1.0.0
torch>=2.0.0,<2.6.0
torchaudio>=2.0.0,<2.6.0
pandas>=2.0.0,<2.3.0
pyyaml>=6.0.1

# Performance & Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.5.0,<2.10.0
pydantic-settings>=2.1.0,<2.7.0

# Logging
loguru>=0.7.2

# HTTP and Async
requests>=2.31.0
aiohttp>=3.9.0

# Data validation
validators>=0.22.0

# Audio processing
ffmpeg-python>=0.2.0

# Web UI
streamlit>=1.28.0
